requirements.txt
streamlit==1.41.1
requests==2.32.3
google-api-python-client==2.86.0
python-dotenv
pandas
fpdf